Split reports into separate files
