+++ docs/reports.rst
@@ -5,18 +5,18 @@ Reports
 
 .. automodule:: robocop.reports
 
-.. autoclass:: robocop.reports.RulesByIdReport
+.. autoclass:: robocop.reports.rules_by_id_report.RulesByIdReport
 
-.. autoclass:: robocop.reports.RulesBySeverityReport
+.. autoclass:: robocop.reports.rules_by_severity_report.RulesBySeverityReport
 
-.. autoclass:: robocop.reports.ReturnStatusReport
+.. autoclass:: robocop.reports.return_status_report.ReturnStatusReport
 
-.. autoclass:: robocop.reports.TimeTakenReport
+.. autoclass:: robocop.reports.time_taken_report.TimeTakenReport
 
-.. autoclass:: robocop.reports.FileStatsReport
+.. autoclass:: robocop.reports.file_stats_report.FileStatsReport
 
-.. autoclass:: robocop.reports.RobocopVersionReport
+.. autoclass:: robocop.reports.robocop_version_report.RobocopVersionReport
 
-.. autoclass:: robocop.reports.TimestampReport
+.. autoclass:: robocop.reports.timestamp_report.TimestampReport
 
-.. autoclass:: robocop.reports.SarifReport
+.. autoclass:: robocop.reports.sarif_report.SarifReport
