+++ robocop/reports/robocop_version_report.py
@@ -1,0 +1,21 @@
+from robocop.reports import Report
+from robocop.version import __version__
+
+
+class RobocopVersionReport(Report):
+    """
+    Report name: ``version``
+
+    Report that returns Robocop version.
+
+    Example::
+
+        Report generated by Robocop version: 2.0.2
+    """
+
+    def __init__(self):
+        self.name = "version"
+        self.description = "Returns Robocop version"
+
+    def get_report(self) -> str:
+        return f"\nReport generated by Robocop version: {__version__}"
